docs/pyproject.toml
[project]
name = "google-docs-mcp-server"
version = "0.1.0"
description = "Obot's Google Docs MCP Server"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp==2.11.2",
    "google-api-python-client>=2.172.0",
]

[project.scripts]
obot-google-docs-mcp = "app.server:streamable_http_server"
obot-google-docs-mcp-stdio = "app.server:stdio_server"

[dependency-groups]
dev = [
    "coverage>=7.9.1",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
    "pytest-cov>=6.2.1",
    "ruff>=0.12.10",
]

[build-system]
requires = ["uv_build>=0.7.21,<0.8.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = "app"
module-root = "."
source-exclude = [
    "tests/",
    ".pytest_cache/",
    "__pycache__/",
    ".venv/",
    "*.pyc",
    "*.pyo",
    "*.pyd",
    ".DS_Store",
]
